#25 refactor: 기능에 따른 이름 변경
