+++ ItNovation/src/main/java/com/ItsTime/ItNovation/domain/mail/EmailAuthRequestDto.java
@@ -1,9 +1,7 @@
-package com.ItsTime.ItNovation.domain.email;
+package com.ItsTime.ItNovation.domain.mail;
 
 
 import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
 
 @Data
 public class EmailAuthRequestDto {
